+++ server/src/main/java/org/opensearch/cluster/metadata/IndexMetadata.java
@@ -96,7 +96,7 @@ import static org.opensearch.cluster.node.DiscoveryNodeFilters.OpType.AND;
 import static org.opensearch.cluster.node.DiscoveryNodeFilters.OpType.OR;
 import static org.opensearch.common.settings.Settings.readSettingsFromStream;
 import static org.opensearch.common.settings.Settings.writeSettingsToStream;
-import static org.opensearch.indices.IndicesService.CLUSTER_DEFAULT_REPLICATION_TYPE_SETTING;
+import static org.opensearch.indices.IndicesService.CLUSTER_REPLICATION_TYPE_SETTING;
 
 /**
  * Index metadata information
@@ -318,10 +318,10 @@ public class IndexMetadata implements Diffable<IndexMetadata>, ToXContentFragmen
 
             @Override
             public void validate(final Boolean value, final Map<Setting<?>, Object> settings) {
-                final Object defaultClusterSettingReplicationType = settings.get(CLUSTER_DEFAULT_REPLICATION_TYPE_SETTING);
+                final Object clusterSettingReplicationType = settings.get(CLUSTER_REPLICATION_TYPE_SETTING);
                 final Object replicationType = settings.get(INDEX_REPLICATION_TYPE_SETTING);
-                if (replicationType != ReplicationType.SEGMENT
-                    && defaultClusterSettingReplicationType != ReplicationType.SEGMENT
+                if ((replicationType).equals(ReplicationType.SEGMENT) == false
+                    && (clusterSettingReplicationType).equals(ReplicationType.SEGMENT) == false
                     && value == true) {
                     throw new IllegalArgumentException(
                         "To enable "
@@ -331,7 +331,7 @@ public class IndexMetadata implements Diffable<IndexMetadata>, ToXContentFragmen
                             + " should be set to "
                             + ReplicationType.SEGMENT
                             + " or "
-                            + CLUSTER_DEFAULT_REPLICATION_TYPE_SETTING.getKey()
+                            + CLUSTER_REPLICATION_TYPE_SETTING.getKey()
                             + " should be set to "
                             + Boolean.TRUE
                     );
@@ -340,7 +340,7 @@ public class IndexMetadata implements Diffable<IndexMetadata>, ToXContentFragmen
 
             @Override
             public Iterator<Setting<?>> settings() {
-                final List<Setting<?>> settings = List.of(INDEX_REPLICATION_TYPE_SETTING, CLUSTER_DEFAULT_REPLICATION_TYPE_SETTING);
+                final List<Setting<?>> settings = List.of(INDEX_REPLICATION_TYPE_SETTING, CLUSTER_REPLICATION_TYPE_SETTING);
                 return settings.iterator();
             }
         },